move faster node load methods from ActiveNode to shared to speed up activerel
