+++ lib/neo4j.rb
@@ -35,6 +35,7 @@ require 'neo4j/shared/validations'
 require 'neo4j/shared/identity'
 require 'neo4j/shared/serialized_properties'
 require 'neo4j/shared/typecaster'
+require 'neo4j/shared/initialize'
 require 'neo4j/shared'
 
 require 'neo4j/active_rel/callbacks'
